Implemented login and signup
