+++ Authorization/Models/LoginModel.cs
@@ -1,0 +1,7 @@
+﻿namespace ShopApp.Authorization.Models;
+
+public class LoginModel
+{
+    public string Username { get; set; }
+    public string Password { get; set; }
+}
