+++ Authorization/Models/SignUpModel.cs
@@ -1,0 +1,8 @@
+﻿namespace ShopApp.Authorization.Models;
+
+public class SignUpModel
+{
+    public string Username { get; set; }
+    public string Password { get; set; }
+    public string Phone { get; set; }
+}
