+++ Authorization/Models/TokenResponse.cs
@@ -1,0 +1,7 @@
+﻿namespace ShopApp.Authorization.Models;
+
+public class TokenResponse
+{
+    public string AccessToken { get; set; }
+    public string RefreshToken { get; set; }
+}
