+++ Authorization/Service/Interfaces/IPasswordService.cs
@@ -1,0 +1,8 @@
+﻿namespace ShopApp.Authorization.Service.Interfaces;
+
+public interface IPasswordService
+{
+    string GenerateSalt();
+    string HashPassword(string password, string salt);
+    bool VerifyPassword(string password, string hashedPassword, string salt);
+}
