+++ Authorization/Service/PasswordService.cs
@@ -1,0 +1,34 @@
+﻿using System.Security.Cryptography;
+using Microsoft.AspNetCore.Cryptography.KeyDerivation;
+using ShopApp.Authorization.Service.Interfaces;
+
+namespace ShopApp.Authorization.Service;
+
+public class PasswordService : IPasswordService
+{
+    public string GenerateSalt()
+    {
+        byte[] salt = new byte[128 / 8];
+        using (var rng = RandomNumberGenerator.Create())
+        {
+            rng.GetBytes(salt);
+        }
+        return Convert.ToBase64String(salt);
+    }
+
+    public string HashPassword(string password, string salt)
+    {
+        return Convert.ToBase64String(KeyDerivation.Pbkdf2(
+            password: password,
+            salt: Convert.FromBase64String(salt),
+            prf: KeyDerivationPrf.HMACSHA256,
+            iterationCount: 10000,
+            numBytesRequested: 256 / 8));
+    }
+
+    public bool VerifyPassword(string password, string hashedPassword, string salt)
+    {
+        var hash = HashPassword(password, salt);
+        return hash == hashedPassword;
+    }
+}
