--- Models/LoginModel.cs
@@ -1,7 +1,0 @@
-﻿namespace ShopApp.Models;
-
-public class LoginModel
-{
-    public string Username { get; set; }
-    public string Password { get; set; }
-}
