--- Models/SignUpModel.cs
@@ -1,8 +1,0 @@
-﻿namespace ShopApp.Models;
-
-public class SignUpModel
-{
-    public string Username { get; set; }
-    public string Password { get; set; }
-    public string Phone { get; set; }
-}
