--- Models/TokenResponse.cs
@@ -1,7 +1,0 @@
-﻿namespace ShopApp.Models;
-
-public class TokenResponse
-{
-    public string AccessToken { get; set; }
-    public string RefreshToken { get; set; }
-}
