+++ Products/Dto/ProductDto.cs
@@ -1,0 +1,6 @@
+﻿namespace ShopApp.Products.Dto
+{
+    public class ProductDto
+    {
+    }
+}
