+++ Products/Enum/ProductEnum.cs
@@ -1,0 +1,36 @@
+﻿namespace ShopApp.Products.Enum;
+
+public class ProductEnums
+{
+    // Enum for product unit
+    public enum ProductUnit
+    {
+        Piece,
+        Pack,
+        Dozen,
+        Gram,
+        Kilogram,
+        Liter,
+        Milliliter
+    }
+
+    // Enum for product status
+    public enum ProductStatus
+    {
+        InStock,
+        OutOfStock,
+        Discontinued
+    }
+
+    // Enum for product categories
+    public enum ProductCategory
+    {
+        Electronics,
+        Books,
+        Clothing,
+        Home,
+        Beauty,
+        Food,
+        Sports
+    }
+}
