+++ Products/Repository/Interface/IProductRepository.cs
@@ -1,0 +1,14 @@
+﻿using ShopApp.Repositories.Interfaces;
+using ShopApp.Products.Entity;
+
+namespace ShopApp.Products.Repository.Interface
+{
+    public interface IProductRepository : IBaseRepository<Product>
+    {
+        Task<IEnumerable<Product>> GetAllProductsAsync();
+        Task<Product> GetProductByIdAsync(Guid productId);
+        Task AddProductAsync(Product product);
+        Task UpdateProductAsync(Product product);
+        Task DeleteProductAsync(Guid productId);
+    }
+}
