+++ Products/Service/Interface/IProductService.cs
@@ -1,0 +1,9 @@
+﻿using ShopApp.Products.Entity;
+
+namespace ShopApp.Products.Service.Interface;
+
+public interface IProductService
+{
+    IEnumerable<Product> SearchProducts(IEnumerable<Product> products, string searchTerm);
+    bool Buy(Guid productId, int quantityToBuy);
+}
