--- Services/Interfaces/IPasswordService.cs
@@ -1,8 +1,0 @@
-﻿namespace ShopApp.Services.Interfaces;
-
-public interface IPasswordService
-{
-    string GenerateSalt();
-    string HashPassword(string password, string salt);
-    bool VerifyPassword(string password, string hashedPassword, string salt);
-}
