--- Services/Interfaces/IProductService.cs
@@ -1,9 +1,0 @@
-﻿using ShopApp.Entities;
-
-namespace ShopApp.Services.Interfaces;
-
-public interface IProductService
-{
-    IEnumerable<Product> SearchProducts(IEnumerable<Product> products, string searchTerm);
-    bool Buy(Guid productId, int quantityToBuy);
-}
